Change 'ctype' field type in the database schema
Cache type ('ctype') field was of text type to be mnemonic, since there
is going to be UI for actually setting it, the mnemonic aspect is now
unnecessary and adding some minor complexity.  Now it will be using
integer, that is established by Grounspeak's SVG file containing the
icons for all the cache types.