Hash passwords stored in backend db
Passwords are now hashed with Argon2 hash before stored in SQLite
database.  The model allows for configurable hashing options, but the
rest of the app does not support that for now, but implementing that
should be easy.  Argon2 is fairly slow, but the default options are
reasonably fast on my slow machine.  Crypt::Passphrase module is used, so
switching to Bcrypt should be trivial, if needed.