Implement displaying of day counts as YyDDD
When day counts (fields 'age' and 'held') are over one year, they will
display as YEARSyDAYS (ie instead of 375 there will be 1y10). This
is implemented as 'age' and 'held' properties now being hashrefs
instead of plain integers; they are utilized by 'daycount' helper
which in turn can be used in the templates.